requirements.txt
ttkbootstrap~=1.10.1
pyserial~=3.5